# server/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BlancBleu API",
    "version": "1.3.0",
    "description": "Plateforme de gestion de transports sanitaires non urgents (Nice 06).\n\n**Authentification** : la plupart des routes exigent soit le cookie httpOnly `bb_access` (émis par /api/auth/login, prioritaire côté browser), soit le header `Authorization: Bearer <jwt>` (SPA / scripts). Les routes service-to-service (training-data, model/retrain) utilisent `X-Service-Token`.",
    "contact": {
      "name": "BlancBleu — Support",
      "email": "support@blancbleu.fr"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Développement local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenu via POST /api/auth/login (utilisable en header Authorization)."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bb_access",
        "description": "Cookie httpOnly émis par /api/auth/login. Utilisé en priorité par les clients browser."
      },
      "serviceTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Service-Token",
        "description": "Token partagé Node ↔ Python pour les routes service-to-service (training-data, model/retrain, model/status)."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Erreur interne du serveur"
          }
        }
      },
...
      "PatientSubdoc": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string",
            "example": "Dupont"
          },
          "prenom": {
            "type": "string",
            "example": "Marie"
          },
          "telephone": {
            "type": "string",
            "example": "0612345678"
          },
          "mobilite": {
            "type": "string",
            "enum": ["ASSIS", "FAUTEUIL_ROULANT", "ALLONGE", "CIVIERE"]
          },
          "oxygene": {
            "type": "boolean",
            "default": false
          },
          "brancardage": {
            "type": "boolean",
            "default": false
          },
          "accompagnateur": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Token manquant, invalide ou expiré",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/transports": {
      "get": {
        "tags": ["Transports"],
        "summary": "Liste paginée des transports",
        "description": "Filtres combinables sur statut, période, véhicule, chauffeur, patient et type. Tri par dateTransport décroissant.",
        "parameters": [
          {
            "in": "query",
            "name": "statut",
            "schema": {
              "type": "string"
            },
            "description": "Filtrer par statut (REQUESTED, ASSIGNED, COMPLETED, …)"
          },
          {
            "in": "query",
            "name": "dateFrom",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page de transports",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transport"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Transports"],
        "summary": "Créer un transport sanitaire",
        "description": "Crée un transport en statut `REQUESTED`. Géocodage automatique des\nadresses (BAN), création du dossier Patient si nécessaire, enqueue\ndu job auto-dispatch BullMQ. Le numéro `TRS-YYYYMMDD-XXXX` est\nattribué atomiquement (Counter Mongo).\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "patient",
                  "typeTransport",
                  "motif",
                  "dateTransport",
                  "heureRDV",
                  "adresseDepart",
                  "adresseDestination"
                ],
                "properties": {
                  "patient": {
                    "$ref": "#/components/schemas/PatientSubdoc"
                  },
                  "typeTransport": {
                    "type": "string",
                    "enum": ["VSL", "TPMR", "AMBULANCE"]
                  },
                  "motif": {
                    "type": "string",
                    "example": "Dialyse"
                  },
                  "dateTransport": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "heureRDV": {
                    "type": "string",
                    "example": "09:30"
                  },
                  "allerRetour": {
                    "type": "boolean",
                    "default": false
                  },
                  "adresseDepart": {
                    "$ref": "#/components/schemas/Adresse"
                  },
                  "adresseDestination": {
                    "$ref": "#/components/schemas/Adresse"
                  },
                  "tauxPriseEnCharge": {
                    "type": "number",
                    "default": 65
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transport créé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transport"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/transports/recurrents": {
      "post": {
        "tags": ["Transports"],
        "summary": "Générer une série de transports récurrents",
        "description": "À partir d'un transport modèle, génère N transports planifiés selon\nune fréquence (hebdo, jours sélectionnés) jusqu'à une date de fin.\nCas d'usage type : dialyse 3×/sem pendant 6 mois.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["modele", "recurrence"],
                "properties": {
                  "modele": {
                    "description": "Payload Transport identique à POST /api/transports",
                    "$ref": "#/components/schemas/Transport"
                  },
                  "recurrence": {
                    "type": "object",
                    "required": ["frequence", "dateFin"],
                    "properties": {
                      "frequence": {
                        "type": "string",
                        "example": "hebdomadaire"
                      },
                      "joursSemaine": {
                        "type": "array",
                        "description": "1=lundi … 7=dimanche",
                        "items": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 7
                        }
                      },
                      "dateFin": {
                        "type": "string",
                        "format": "date"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Série créée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer",
                      "example": 24
                    },
                    "transports": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transport"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/transports/{id}": {
      "get": {
        "tags": ["Transports"],
        "summary": "Détail d'un transport",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transport trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transport"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": ["Transports"],
        "summary": "Modifier un transport",
        "description": "Whitelist de champs modifiables : `notes`, `heureDepart`, `allerRetour`,\n`adresseDepart`, `adresseDestination`, `tauxPriseEnCharge`. Les autres\nchamps (statut, vehicule, chauffeur) passent par des transitions dédiées.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "notes": {
                    "type": "string"
                  },
                  "heureDepart": {
                    "type": "string",
                    "example": "09:00"
                  },
                  "allerRetour": {
                    "type": "boolean"
                  },
                  "adresseDepart": {
                    "$ref": "#/components/schemas/Adresse"
                  },
                  "adresseDestination": {
                    "$ref": "#/components/schemas/Adresse"
                  },
                  "tauxPriseEnCharge": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transport mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transport"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/transports/{id}/assign": {
      "patch": {
        "tags": ["Transports"],
        "summary": "Assigner véhicule + chauffeur à un transport",
        "description": "Transition `CONFIRMED/SCHEDULED → ASSIGNED`. Claim atomique du\nvéhicule (`Vehicle.findOneAndUpdate` garde `statut=Disponible`),\nenqueue push FCM canal critique vers le chauffeur. RBAC : dispatcher\n/ superviseur / admin.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["vehiculeId"],
                "properties": {
                  "vehiculeId": {
                    "type": "string"
                  },
                  "chauffeurId": {
                    "type": "string",
                    "nullable": true,
                    "description": "Si null, dérivé du shift actif du véhicule"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transport assigné",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transport"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/transports/{id}/complete": {
      "patch": {
        "tags": ["Transports"],
        "summary": "Marquer transport comme terminé (transition de statut représentative)",
        "description": "Transition `ARRIVED_AT_DESTINATION → COMPLETED`. Libère le véhicule\n(`statut=Disponible`), enqueue job billing BullMQ, calcule la durée\nréelle, émet WebSocket `transport:status_updated`. Endpoint emblématique\nde la state machine — les autres transitions suivent le même contrat\n(404 si introuvable, 409 si transition interdite, 200 + Transport\nmis à jour). Voir transitions sœurs : `/cancel`, `/no-show`, `/en-route`,\n`/arrived`, `/on-board`, `/destination`, etc.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dureeAttenteMinutes": {
                    "type": "integer",
                    "description": "Saisie chauffeur"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transport terminé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transport"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
  }
}